+++ packages/eth-memento/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v2.0.2
 
 - Add compatibility with @alethio/cms v1.0.0-beta.10